+++ mlflow/store/artifact/repo/azure_blob.py
@@ -6,7 +6,7 @@ from six.moves import urllib
 
 from mlflow.entities import FileInfo
 from mlflow.exceptions import MlflowException
-from mlflow.store.artifact.artifact_repo import ArtifactRepository
+from mlflow.store.artifact.repo import ArtifactRepository
 
 
 class AzureBlobArtifactRepository(ArtifactRepository):
