+++ mlflow/store/artifact/repo/ftp.py
@@ -7,7 +7,7 @@ import posixpath
 from six.moves import urllib
 
 from mlflow.entities.file_info import FileInfo
-from mlflow.store.artifact.artifact_repo import ArtifactRepository
+from mlflow.store.artifact.repo import ArtifactRepository
 from mlflow.utils.file_utils import relative_path_to_artifact_path
