+++ mlflow/store/artifact/repo/gcs.py
@@ -4,7 +4,7 @@ import posixpath
 from six.moves import urllib
 
 from mlflow.entities import FileInfo
-from mlflow.store.artifact.artifact_repo import ArtifactRepository
+from mlflow.store.artifact.repo import ArtifactRepository
 from mlflow.utils.file_utils import relative_path_to_artifact_path
