+++ mlflow/store/artifact/repo/registry.py
@@ -2,16 +2,16 @@ import entrypoints
 import warnings
 
 from mlflow.exceptions import MlflowException
-from mlflow.store.artifact.azure_blob_artifact_repo import AzureBlobArtifactRepository
-from mlflow.store.artifact.dbfs_artifact_repo import dbfs_artifact_repo_factory
-from mlflow.store.artifact.ftp_artifact_repo import FTPArtifactRepository
-from mlflow.store.artifact.gcs_artifact_repo import GCSArtifactRepository
-from mlflow.store.artifact.hdfs_artifact_repo import HdfsArtifactRepository
-from mlflow.store.artifact.local_artifact_repo import LocalArtifactRepository
-from mlflow.store.artifact.models_artifact_repo import ModelsArtifactRepository
-from mlflow.store.artifact.runs_artifact_repo import RunsArtifactRepository
-from mlflow.store.artifact.s3_artifact_repo import S3ArtifactRepository
-from mlflow.store.artifact.sftp_artifact_repo import SFTPArtifactRepository
+from mlflow.store.artifact.repo.azure_blob import AzureBlobArtifactRepository
+from mlflow.store.artifact.repo.dbfs import dbfs_artifact_repo_factory
+from mlflow.store.artifact.repo.ftp import FTPArtifactRepository
+from mlflow.store.artifact.repo.gcs import GCSArtifactRepository
+from mlflow.store.artifact.repo.hdfs import HdfsArtifactRepository
+from mlflow.store.artifact.repo.local import LocalArtifactRepository
+from mlflow.store.artifact.repo.models import ModelsArtifactRepository
+from mlflow.store.artifact.repo.runs import RunsArtifactRepository
+from mlflow.store.artifact.repo.s3 import S3ArtifactRepository
+from mlflow.store.artifact.repo.sftp import SFTPArtifactRepository
 
 from mlflow.utils.uri import get_uri_scheme
