+++ mlflow/tracking/_model_registry/fluent.py
@@ -1,6 +1,6 @@
 from mlflow.exceptions import MlflowException
 from mlflow.protos.databricks_pb2 import RESOURCE_ALREADY_EXISTS, ErrorCode
-from mlflow.store.artifact.runs_artifact_repo import RunsArtifactRepository
+from mlflow.store.artifact.repo.runs import RunsArtifactRepository
 from mlflow.tracking import MlflowClient
 from mlflow.utils import experimental
 from mlflow.utils.logging_utils import eprint
