+++ mlflow/tracking/_tracking_service/client.py
@@ -13,7 +13,7 @@ from mlflow.tracking._tracking_service import utils
 from mlflow.utils.validation import _validate_param_name, _validate_tag_name, _validate_run_id, \
     _validate_experiment_artifact_location, _validate_experiment_name, _validate_metric
 from mlflow.entities import Param, Metric, RunStatus, RunTag, ViewType, ExperimentTag
-from mlflow.store.artifact.artifact_repository_registry import get_artifact_repository
+from mlflow.store.artifact.repo.registry import get_artifact_repository
 from mlflow.utils.mlflow_tags import MLFLOW_USER
