+++ tests/pyfunc/test_model_export_with_class_and_artifacts.py
@@ -22,7 +22,7 @@ import mlflow.pyfunc.scoring_server as pyfunc_scoring_server
 import mlflow.sklearn
 from mlflow.exceptions import MlflowException
 from mlflow.models import Model
-from mlflow.store.artifact.s3_artifact_repo import S3ArtifactRepository
+from mlflow.store.artifact.repo.s3 import S3ArtifactRepository
 from mlflow.tracking.artifact_utils import get_artifact_uri as utils_get_artifact_uri, \
     _download_artifact_from_uri
 from mlflow.utils.environment import _mlflow_conda_env
