+++ tests/store/artifact/test_artifact_repo.py
@@ -2,7 +2,7 @@ import mock
 import pytest
 
 from mlflow.entities import FileInfo
-from mlflow.store.artifact.artifact_repo import ArtifactRepository
+from mlflow.store.artifact.repo import ArtifactRepository
 
 
 class ArtifactRepositoryImpl(ArtifactRepository):
