+++ tests/store/artifact/test_dbfs_artifact_repo_delegation.py
@@ -2,16 +2,16 @@ import os
 import mock
 import pytest
 
-from mlflow.store.artifact.artifact_repository_registry import get_artifact_repository
-from mlflow.store.artifact.local_artifact_repo import LocalArtifactRepository
-from mlflow.store.artifact.dbfs_artifact_repo import DbfsRestArtifactRepository
+from mlflow.store.artifact.repo.registry import get_artifact_repository
+from mlflow.store.artifact.repo.local import LocalArtifactRepository
+from mlflow.store.artifact.repo.dbfs import DbfsRestArtifactRepository
 
 from mlflow.utils.rest_utils import MlflowHostCreds
 
 
 @pytest.fixture()
 def host_creds_mock():
-    with mock.patch('mlflow.store.artifact.dbfs_artifact_repo._get_host_creds_from_default_store') \
+    with mock.patch('mlflow.store.artifact.repo.dbfs._get_host_creds_from_default_store') \
             as get_creds_mock:
         get_creds_mock.return_value = lambda: MlflowHostCreds('http://host')
         yield
