+++ tests/store/artifact/test_hdfs_artifact_repo.py
@@ -8,8 +8,10 @@ from mock import call, mock_open
 from pyarrow import HadoopFileSystem
 
 from mlflow.entities import FileInfo
-from mlflow.store.artifact.hdfs_artifact_repo import HdfsArtifactRepository, _resolve_base_path, \
-    _relative_path_remote, _parse_extra_conf, _download_hdfs_file
+from mlflow.store.artifact.repo.hdfs import (
+    HdfsArtifactRepository, _resolve_base_path, _relative_path_remote,
+    _parse_extra_conf, _download_hdfs_file,
+)
 from mlflow.utils.file_utils import TempDir
