+++ tests/store/artifact/test_sftp_artifact_repo.py
@@ -2,8 +2,8 @@ from mock import MagicMock
 import pytest
 from tempfile import NamedTemporaryFile
 import pysftp
-from mlflow.store.artifact.artifact_repository_registry import get_artifact_repository
-from mlflow.store.artifact.sftp_artifact_repo import SFTPArtifactRepository
+from mlflow.store.artifact.repo.registry import get_artifact_repository
+from mlflow.store.artifact.repo.sftp import SFTPArtifactRepository
 from mlflow.utils.file_utils import TempDir
 import os
 import posixpath
